Cast buffer sizes to u32 in example code

+++ examples/feature_showcase/source/basic_effect_test/basic_effect_test.cpp
@@ -71,7 +71,7 @@ std::unique_ptr<Error> BasicEffectTest::initialize()
 
         std::tie(vertexBuffer1, err) = graphicsDevice->createVertexBuffer(
             verticesCombo.data(),
-            verticesCombo.size(),
+            static_cast<u32>(verticesCombo.size()),
             sizeof(VertexCombined),
             gpu::BufferUsage::Immutable);
 
@@ -117,7 +117,7 @@ std::unique_ptr<Error> BasicEffectTest::initialize()
 
         std::tie(vertexBuffer2, err) = graphicsDevice->createVertexBuffer(
             verticesCombo.data(),
-            verticesCombo.size(),
+            static_cast<u32>(verticesCombo.size()),
             sizeof(VertexCombined),
             gpu::BufferUsage::Immutable);
 
@@ -147,7 +147,7 @@ std::unique_ptr<Error> BasicEffectTest::initialize()
         std::tie(indexBuffer, err) = graphicsDevice->createIndexBuffer(
             gpu::IndexFormat::UInt16,
             indices.data(),
-            indices.size(),
+            static_cast<u32>(indices.size()),
             gpu::BufferUsage::Immutable);
 
         if (err != nullptr) {
