+++ examples/feature_showcase/source/hardware_instancing_test/hardware_instancing_test.cpp
@@ -51,7 +51,7 @@ std::unique_ptr<Error> HardwareInstancingTest::initialize()
 
         std::tie(vertexBuffer, err) = graphicsDevice->createVertexBuffer(
             verticesCombo.data(),
-            verticesCombo.size(),
+            static_cast<u32>(verticesCombo.size()),
             sizeof(VertexCombined),
             gpu::BufferUsage::Immutable);
 
@@ -66,7 +66,7 @@ std::unique_ptr<Error> HardwareInstancingTest::initialize()
         std::tie(indexBuffer, err) = graphicsDevice->createIndexBuffer(
             gpu::IndexFormat::UInt16,
             indices.data(),
-            indices.size(),
+            static_cast<u32>(indices.size()),
             gpu::BufferUsage::Immutable);
 
         if (err != nullptr) {
