+++ examples/quickstart/source/game_main.cpp
@@ -59,7 +59,7 @@ std::unique_ptr<Error> GameMain::initialize()
 
         std::tie(vertexBuffer, err) = graphicsDevice->createVertexBuffer(
             verticesCombo.data(),
-            verticesCombo.size(),
+            static_cast<u32>(verticesCombo.size()),
             sizeof(VertexCombined),
             gpu::BufferUsage::Immutable);
 
@@ -74,7 +74,7 @@ std::unique_ptr<Error> GameMain::initialize()
         std::tie(indexBuffer, err) = graphicsDevice->createIndexBuffer(
             gpu::IndexFormat::UInt16,
             indices.data(),
-            indices.size(),
+            static_cast<u32>(indices.size()),
             gpu::BufferUsage::Immutable);
 
         if (err != nullptr) {
